Fix bug in convertEntryStringDatestoDates
Isssue: the member passed to new Date() was converted to a Date object
too.

Cause: in line 7 before this commit, a new object was created, assigning
the value of the argument passed to the function (entry). Apparently,
this assignment wasn't copying all the values and properties of 'entry'
to the new variable, but making the variable refer to 'entry'. This way,
'entry.start' and 'entry.end' types were changing from string to Date
together with 'newEntry.start' and 'newEntry.end' values.

Solution: use 'Object.assign()' method to copy all the values of 'entry'
into 'newEntry'.