Update psatEntry.js tests to be more maintainable
Each step in the sequence  stored in a single object with 'actual' and
'expected' keys:
  - 'actual' key values obtained applying the functions to test on
    'actual' values obtained in previous steps.
  - 'expected' key values hardcoded.

This update was motivated by the fact that is not practical to update
the expected values if a new member is added to pastData.json entries.
It was necessary to optimize the cases to test, otherwise the
maintainability of the test suit would be unnecessarily painful.